Move MyEvaluator into the mydsl package
